+++ resources/views/components/logout-link.blade.php
@@ -1,6 +1,6 @@
 @props(['useIcons' => false])
 
-<form method="POST" action="{{ route('logout') }}">
+<form method="POST" action="{{ route('logout') }}" class="my-0">
     @csrf
     <a href="{{ route('logout') }}" onclick="event.preventDefault(); this.closest('form').submit();">
         @if($useIcons)