+++ src/config/authit.php
@@ -3,6 +3,12 @@
 return [
 
     /**
+     * This configuration option allows you to choose between a single `name`
+     * field or a `firstname` and `lastname` field when the user registers.
+     */
+    'use_single_name_field' => env('NK_USE_SINGLE_NAME_FIELD', true),
+
+    /**
      * ----------------------------------------------------------------------
      * Registration Routes
      * ----------------------------------------------------------------------