+++ src/database/migrations/add_authit_columns_to_users_table.php
@@ -6,20 +6,35 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-
     public function up()
     {
-        if (!Schema::hasColumn('users', 'avatar')) {
-            Schema::table('users', function (Blueprint $table) {
-                $table->string('avatar', 2048)->nullable();
-            });
-        }
+        Schema::table('users', function (Blueprint $table) {
+            if (!config('authit.use_single_name_field')) {
+                $table->dropColumn('name');
+                $table->string('firstname', 128)->nullable()->after('id');
+                $table->string('lastname', 128)->nullable()->after('firstname');
+            }
+            $table->string('avatar', 2048)->nullable();
+        });
     }
 
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('avatar');
+            if (Schema::hasColumn('users', 'avatar')) {
+                $table->dropColumn('avatar');
+            }
+            if (!config('authit.use_single_name_field')) {
+                if (Schema::hasColumn('users', 'firstname')) {
+                    $table->dropColumn('firstname');
+                }
+                if (Schema::hasColumn('users', 'lastname')) {
+                    $table->dropColumn('lastname');
+                }
+                if (!Schema::hasColumn('users', 'name')) {
+                    $table->string('name', 255)->nullable()->after('id');
+                }
+            }
         });
     }
 };